Update from latest duplicati/duplicati

+++ guiTests/guiTest.py
@@ -8,10 +8,11 @@ from selenium import webdriver
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions
+from selenium.webdriver.firefox.options import Options
 
 if "TRAVIS_BUILD_NUMBER" in os.environ:
     if "SAUCE_USERNAME" not in os.environ:
-        print "No sauce labs login credentials found. Stopping tests..."
+        print("No sauce labs login credentials found. Stopping tests...")
         sys.exit(0)
 
     capabilities = {'browserName': "firefox"}
@@ -28,11 +29,12 @@ if "TRAVIS_BUILD_NUMBER" in os.environ:
     driver = webdriver.Remote(command_executor="http://%s/wd/hub" % hub_url, desired_capabilities=capabilities)
 else:
     # local
-    print "Using LOCAL webdriver"
+    print("Using LOCAL webdriver")
     profile = webdriver.FirefoxProfile()
     profile.set_preference("intl.accept_languages", "en")
-    driver = webdriver.Firefox(profile)
-    driver.maximize_window()
+    options = Options()
+    options.set_headless(headless=True)
+    driver = webdriver.Firefox(profile, firefox_options=options)
 
 
 def write_random_file(size, filename):
@@ -87,6 +89,7 @@ DIRECT_RESTORE_FOLDER = os.path.abspath("duplicati_gui_test_direct_restore")
 time.sleep(5)
 
 driver.implicitly_wait(10)
+driver.maximize_window()
 driver.get("http://localhost:8200/ngax/index.html")
 
 if "Duplicati" not in driver.title:
@@ -139,10 +142,10 @@ driver.find_element_by_link_text(BACKUP_NAME).click()
 wait_for_text(60, "//div[@class='task ng-scope']/dl[2]/dd[1]", "(took ")
 
 # Restore
-if len([n for n in driver.find_elements_by_xpath("//span[contains(text(),'Restore files ...')]") if n.is_displayed()]) == 0:
+if len([n for n in driver.find_elements_by_xpath(u"//span[contains(text(),'Restore files \u2026')]") if n.is_displayed()]) == 0:
     driver.find_element_by_link_text(BACKUP_NAME).click()
 
-[n for n in driver.find_elements_by_xpath("//span[contains(text(),'Restore files ...')]") if n.is_displayed()][0].click()
+[n for n in driver.find_elements_by_xpath(u"//span[contains(text(),'Restore files \u2026')]") if n.is_displayed()][0].click()
 driver.find_element_by_xpath("//span[contains(text(),'" + SOURCE_FOLDER + "')]")  # wait for filelist
 time.sleep(1)
 driver.find_element_by_xpath("//restore-file-picker/ul/li/div/a[2]").click()  # select root folder checkbox
@@ -159,8 +162,10 @@ wait_for_text(60, "//form[@id='restore']/div[3]/h3/div[1]", "Your files and fold
 sha1_restore = sha1_folder(RESTORE_FOLDER)
 
 # cleanup: delete source and restore folder and rename destination folder for direct restore
-shutil.rmtree(SOURCE_FOLDER)
-shutil.rmtree(RESTORE_FOLDER)
+if os.path.exists(SOURCE_FOLDER):
+    shutil.rmtree(SOURCE_FOLDER)
+if os.path.exists(RESTORE_FOLDER):
+    shutil.rmtree(RESTORE_FOLDER)
 os.rename(DESTINATION_FOLDER, DESTINATION_FOLDER_DIRECT_RESTORE)
 
 # direct restore
@@ -193,9 +198,9 @@ wait_for_text(60, "//form[@id='restore']/div[3]/h3/div[1]", "Your files and fold
 # hash direct restore files
 sha1_direct_restore = sha1_folder(DIRECT_RESTORE_FOLDER)
 
-print "Source hashes: " + str(sha1_source)
-print "Restore hashes: " + str(sha1_restore)
-print "Direct Restore hashes: " + str(sha1_direct_restore)
+print("Source hashes: " + str(sha1_source))
+print("Restore hashes: " + str(sha1_restore))
+print("Direct Restore hashes: " + str(sha1_direct_restore))
 
 # Tell Sauce Labs to stop the test
 driver.quit()
